Move axiom validation tests to ristretto_specs.rs, revert window.rs trigger changes
- Move test_ristretto_decode_axioms and pow_field_element from ristretto.rs
  to specs/ristretto_specs.rs, co-locating tests with axiom definitions.
- Revert unintentional #[trigger] -> #![auto] changes in window.rs
  introduced by verusfmt.

Made-with: Cursor
