+++ curve25519-dalek/src/window.rs
@@ -484,8 +484,7 @@ impl<'a> From<&'a EdwardsPoint> for LookupTable<ProjectiveNielsPoint> {
         // Limb bounds for all entries built so far
 
                 forall|k: int|
-                    #![auto]
-                    0 <= k <= j ==> fe51_limbs_bounded(&points[k].Y_plus_X, 54)
+                    0 <= k <= j ==> fe51_limbs_bounded(&(#[trigger] points[k]).Y_plus_X, 54)
                         && fe51_limbs_bounded(&points[k].Y_minus_X, 54) && fe51_limbs_bounded(
                         &points[k].Z,
                         54,
@@ -604,8 +603,7 @@ impl<'a> From<&'a EdwardsPoint> for LookupTable<AffineNielsPoint> {
         for j in 0..7
             invariant
                 forall|k: int|
-                    #![auto]
-                    0 <= k <= j ==> fe51_limbs_bounded(&points[k].y_plus_x, 54)
+                    0 <= k <= j ==> fe51_limbs_bounded(&(#[trigger] points[k]).y_plus_x, 54)
                         && fe51_limbs_bounded(&points[k].y_minus_x, 54) && fe51_limbs_bounded(
                         &points[k].xy2d,
                         54,
@@ -823,11 +821,11 @@ impl<'a> From<&'a EdwardsPoint> for NafLookupTable5<ProjectiveNielsPoint> {
         // Limb bounds for all entries built so far
 
                 forall|k: int|
-                    #![auto]
-                    0 <= k <= i ==> fe51_limbs_bounded(&Ai[k].Y_plus_X, 54) && fe51_limbs_bounded(
-                        &Ai[k].Y_minus_X,
+                    0 <= k <= i ==> fe51_limbs_bounded(&(#[trigger] Ai[k]).Y_plus_X, 54)
+                        && fe51_limbs_bounded(&Ai[k].Y_minus_X, 54) && fe51_limbs_bounded(
+                        &Ai[k].Z,
                         54,
-                    ) && fe51_limbs_bounded(&Ai[k].Z, 54) && fe51_limbs_bounded(&Ai[k].T2d, 54),
+                    ) && fe51_limbs_bounded(&Ai[k].T2d, 54),
                 // Validity for all entries built so far
                 forall|k: int| 0 <= k <= i ==> is_valid_projective_niels_point(#[trigger] Ai[k]),
                 // Mathematical correspondence: entry[k] represents (2*k+1)*A
@@ -923,11 +921,11 @@ impl<'a> From<&'a EdwardsPoint> for NafLookupTable5<AffineNielsPoint> {
         for i in 0..7
             invariant
                 forall|k: int|
-                    #![auto]
-                    0 <= k <= i ==> fe51_limbs_bounded(&Ai[k].y_plus_x, 54) && fe51_limbs_bounded(
-                        &Ai[k].y_minus_x,
+                    0 <= k <= i ==> fe51_limbs_bounded(&(#[trigger] Ai[k]).y_plus_x, 54)
+                        && fe51_limbs_bounded(&Ai[k].y_minus_x, 54) && fe51_limbs_bounded(
+                        &Ai[k].xy2d,
                         54,
-                    ) && fe51_limbs_bounded(&Ai[k].xy2d, 54),
+                    ),
                 forall|k: int| 0 <= k <= i ==> is_valid_affine_niels_point(#[trigger] Ai[k]),
                 forall|k: int|
                     0 <= k <= i ==> affine_niels_point_as_affine_edwards(#[trigger] Ai[k])
@@ -1115,11 +1113,11 @@ impl<'a> From<&'a EdwardsPoint> for NafLookupTable8<ProjectiveNielsPoint> {
         for i in 0..63
             invariant
                 forall|k: int|
-                    #![auto]
-                    0 <= k <= i ==> fe51_limbs_bounded(&Ai[k].Y_plus_X, 54) && fe51_limbs_bounded(
-                        &Ai[k].Y_minus_X,
+                    0 <= k <= i ==> fe51_limbs_bounded(&(#[trigger] Ai[k]).Y_plus_X, 54)
+                        && fe51_limbs_bounded(&Ai[k].Y_minus_X, 54) && fe51_limbs_bounded(
+                        &Ai[k].Z,
                         54,
-                    ) && fe51_limbs_bounded(&Ai[k].Z, 54) && fe51_limbs_bounded(&Ai[k].T2d, 54),
+                    ) && fe51_limbs_bounded(&Ai[k].T2d, 54),
                 forall|k: int| 0 <= k <= i ==> is_valid_projective_niels_point(#[trigger] Ai[k]),
                 forall|k: int|
                     0 <= k <= i ==> projective_niels_point_as_affine_edwards(#[trigger] Ai[k])
@@ -1216,11 +1214,11 @@ impl<'a> From<&'a EdwardsPoint> for NafLookupTable8<AffineNielsPoint> {
         for i in 0..63
             invariant
                 forall|k: int|
-                    #![auto]
-                    0 <= k <= i ==> fe51_limbs_bounded(&Ai[k].y_plus_x, 54) && fe51_limbs_bounded(
-                        &Ai[k].y_minus_x,
+                    0 <= k <= i ==> fe51_limbs_bounded(&(#[trigger] Ai[k]).y_plus_x, 54)
+                        && fe51_limbs_bounded(&Ai[k].y_minus_x, 54) && fe51_limbs_bounded(
+                        &Ai[k].xy2d,
                         54,
-                    ) && fe51_limbs_bounded(&Ai[k].xy2d, 54),
+                    ),
                 forall|k: int| 0 <= k <= i ==> is_valid_affine_niels_point(#[trigger] Ai[k]),
                 forall|k: int|
                     0 <= k <= i ==> affine_niels_point_as_affine_edwards(#[trigger] Ai[k])
